chore(cli): fix export of metrics and traces
