+++ .local.env
@@ -37,6 +37,7 @@ PIPELINE_URI=http://pipeline.cube-creator.lndo.site
 OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://otel.cube-creator.lndo.site/v1/traces
 OTEL_EXPORTER_OTLP_METRICS_ENDPOINT=http://otel.cube-creator.lndo.site/v1/metrics
 OTEL_METRICS_EXPORTER=otlp
+OTEL_LOG_LEVEL=debug
 
 # Shared Dimensions
 MANAGED_DIMENSIONS_GRAPH=https://lindas.admin.ch/cube/dimension
