+++ otel/collector.yaml
@@ -2,7 +2,10 @@ receivers:
   otlp:
     protocols:
       grpc:
+        endpoint: 0.0.0.0:4317
+        max_recv_msg_size_mib: 100
       http:
+        endpoint: 0.0.0.0:55681
 
 exporters:
   jaeger:
@@ -13,6 +16,7 @@ exporters:
     endpoint: "0.0.0.0:8889"
     resource_to_telemetry_conversion:
       enabled: true
+  logging:
 
 processors:
   batch:
@@ -26,4 +30,4 @@ service:
     metrics:
       receivers: [otlp]
       processors: [batch]
-      exporters: [prometheus]
+      exporters: [prometheus, logging]
